+++ Assets/CommonCore/Core/Util/StringUtils.cs
@@ -15,5 +15,22 @@ namespace CommonCore
         {
             return strA?.IndexOf(strB, comparisonType) >= 0;
         }
+
+        /// <summary>
+        /// Counts occurrences of a character in a string
+        /// </summary>
+        public static int CountChar(this string str, char ch)
+        {
+            //according to https://stackoverflow.com/questions/541954/how-would-you-count-occurrences-of-a-string-actually-a-char-within-a-string this is the fastest way
+
+            int count = 0;
+            for(int i = 0; i < str.Length; i++)
+            {
+                if (str[i] == ch)
+                    count++;
+            }
+
+            return count;                
+        }
     }
 }